Refactored structure of assignments hook
